Simplify the transport & move parsing logic into protocol
