+++ packages/thrift-server-core/src/transports/Transport.ts
@@ -1,30 +1,18 @@
 export interface ITransport {
-  commitPosition(): void
-  rollbackPosition(): void
+  open(): void
+  close(): void
 
   read(size: number): Promise<Buffer>
-  readByte(): Promise<number>
-  readBool(): Promise<boolean>
-  readI16(): Promise<number>
-  readI32(): Promise<number>
-  readDouble(): Promise<number>
-  readString(size: number): Promise<string>
 
-  write(buf: Buffer | string): void
-  writeByte(value: number): void
-  writeBool(value: boolean): void
-  writeI16(value: number): void
-  writeI32(value: number): void
-  writeDouble(value: number): void
-  writeString(value: string): void
+  // write(buf: Buffer | string): void
+  write(buf: Buffer): void
 
   flush(): void
 
+  // TODO: Below methods might not be needed
   consume(bytesConsumed: number): void
-
   borrow(): { buf: Buffer; readIndex: number; writeIndex: number; }
-
+  commitPosition(): void
+  rollbackPosition(): void
   isOpen(): boolean
-  open(): void
-  close(): void
 }
